+++ Source/EventFlow.EventStores.EventStore/Subscriptions/StreamName.cs
@@ -1,0 +1,46 @@
+// The MIT License (MIT)
+//
+// Copyright (c) 2015-2020 Rasmus Mikkelsen
+// Copyright (c) 2015-2020 eBay Software Foundation
+// https://github.com/eventflow/EventFlow
+//
+// Permission is hereby granted, free of charge, to any person obtaining a copy of
+// this software and associated documentation files (the "Software"), to deal in
+// the Software without restriction, including without limitation the rights to
+// use, copy, modify, merge, publish, distribute, sublicense, and/or sell copies of
+// the Software, and to permit persons to whom the Software is furnished to do so,
+// subject to the following conditions:
+//
+// The above copyright notice and this permission notice shall be included in all
+// copies or substantial portions of the Software.
+//
+// THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
+// IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY, FITNESS
+// FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE AUTHORS OR
+// COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER LIABILITY, WHETHER
+// IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN
+// CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
+
+namespace EventFlow.EventStores.EventStore.Subscriptions
+{
+    public class StreamName
+    {
+        string Value { get; }
+
+        const string AllStreamName = "$all";
+
+        StreamName(string value) => Value = value;
+
+        public static StreamName AllStream => new StreamName(AllStreamName);
+
+        public static StreamName For<T>(string id) => new StreamName($"{typeof(T).Name}-{id}");
+
+        public static StreamName Custom(string streamName) => new StreamName(streamName);
+
+        public bool IsAllStream => Value.Equals(AllStreamName);
+
+        public override string ToString() => Value ?? "";
+
+        public static implicit operator string(StreamName self) => self.ToString();
+    }
+}