+++ RELEASE_NOTES.md
@@ -1,6 +1,9 @@
 ### New in 0.81 (not released yet)
 
 * Fixed: You can now create `Id : Identity<Id>`
+* EventStore has been updated to Nuget Library 20.6.1
+  and a UseEventStoreSubscriptions extensions has been added to help subscribe to stream
+  events based on a filter and then publish them to your internal microservices (pubsub)
 
 ### New in 0.80.4377 (released 2020-10-01)